refactor ClaimForm
