+++ src/components/forms/claims/_components/LossReasonRadioOptions.svelte
@@ -1,0 +1,16 @@
+<script lang="ts">
+import RadioOptions from '../../../RadioOptions.svelte'
+import { claimIncidentTypes } from 'data/types/claim-incident-types'
+
+export let lossReason: string
+
+let lossReasonOptions: { label: string; value: string }[] = []
+
+// Calculate dynamic options for radio-button prompts.
+$: lossReasonOptions = $claimIncidentTypes.map(({ name, description }) => ({ label: name, value: name, description }))
+</script>
+
+<p>
+  <span class="header">Reason for loss or damage</span>
+  <RadioOptions name="lossReason" options={lossReasonOptions} bind:value={lossReason} />
+</p>
