+++ src/components/forms/claims/_components/index.ts
@@ -1,0 +1,5 @@
+import LossReasonRadioOptions from './LossReasonRadioOptions.svelte'
+import PayoutRadioOptions from './PayoutRadioOptions.svelte'
+import RepairableRadioOptions from './RepairableRadioOptions.svelte'
+
+export { LossReasonRadioOptions, PayoutRadioOptions, RepairableRadioOptions }
